@@ -25,6 +25,7 @@
   },
   "homepage": "https://github.com/ChenHao1995/chenhao-react#readme",
   "devDependencies": {
+    "antd": "^3.16.2",
     "antd-mobile": "^2.1.8",
     "babel": "^6.23.0",
     "babel-cli": "^6.26.0",
@@ -58,6 +59,7 @@
     "history": "^4.7.2",
     "html-loader": "^0.5.5",
     "html-webpack-plugin": "^2.30.1",
+    "immutability-helper": "^3.0.0",
     "immutable": "^3.8.2",
     "inquirer": "^6.2.2",
     "isomorphic-fetch": "^2.2.1",
@@ -67,6 +69,8 @@
     "nodemon": "^1.14.11",
     "open": "0.0.5",
     "react": "^16.2.0",
+    "react-dnd": "^7.4.5",
+    "react-dnd-html5-backend": "^7.4.4",
     "react-dom": "^16.2.0",
     "react-redux": "^5.0.7",
     "react-router": "^4.2.0",
@@ -74,10 +78,12 @@
     "react-router-redux": "^5.0.0-alpha.9",
     "redux": "^3.7.2",
     "redux-saga": "^0.16.2",
+    "source-map-loader": "^0.2.4",
     "style-loader": "^0.20.3",
+    "ts-loader": "^3.5.0",
     "typescript": "^3.3.3333",
     "url-loader": "^1.0.1",
-    "webpack": "^3.10.0",
+    "webpack": "^3.12.0",
     "webpack-dev-server": "^2.11.1",
     "webpack-hot-middleware": "^2.24.3",
     "webpack-node-externals": "^1.7.2",
